requirements.txt
cloudpickle==2.0.0
ipython==8.8.0
laspy==2.0.3
matplotlib==3.6.2
numpy==1.23.5
open3d==0.10.0.0
pandas==1.1.5
Pillow==9.4.0
pyntcloud==0.1.6
PyYAML==6.0
scikit_learn==1.2.1
scipy==1.6.1
seaborn==0.11.2
sympy==1.10.1
sympytorch==0.1.1
torch==1.13.0
torchmetrics==0.9.0
torchsummary==1.5.1
torchvision==0.11.1+cu102
torchviz==0.0.2
tqdm==4.62.3
webcolors==1.12